RHDM-961 Random test failures (#2352)
The error message indicated that sometimes the format function being
called by an accumulate, in the rule "processDerivedField_MapValues",
was being called with a "type" parameter that was not a String. I
updated the function's signature to take a generic Object for the "type"
parameter, and then modified the function to check to see if the
parameter was indeed a String. If it is a String then the function works
as it was originally coded; otherwise it returns the passed in "val"
parameter, just as though the "type" parameter was null.