Groupby - Interpreted DRL (#5292)
* WIP: add wiring and expression compilation code for groupby

Currently the group key and accumulation pair is calculated correctly,
but constraints down the stream gets Object[] containing the pair
($sumOfAge, $initial) instead of what it expected ($sumOfAge by itself).

* Use ArrayReader for reading groupby key and result

The confusing messages about Object[] cannot be cast to Integer
was because SelfClassFieldReader was being used to read the
accumulate result, which is an Object[] when groupby is used.
The key and result declarations now use an ArrayReader to get
their value (which use SelfClassFieldReader to get the Object
array).

* Use correct declarations for groupping function

Previously, the code used the same declarations as the
accumulator, which leads to errors when the grouping function
uses different variables than the accumulators.

Work still need to be done:

- Allow group key to be used as a variable in accumulators
- Figure out why Pair.create cannot be analyzed in strict-mode
- Find out cause of ClassCastException in testGroupByUpdatingKey

* Allow groupby keys to be used in accumulate functions

- For some reason the grouping key is not put into BoundIdentifiers
despite being in the compiled function, being accessible in context,
and being present in the declCls map, meaning we need to check if
the parameter contains the grouping key, and if so, add it to the
usedIdentifiers.

- To allow accumulates to access the grouping key, the grouping function
is compiled first, which calls a supplier to generate the inner
accumulate when the groupby key declaration is known.

* Fix off-by-one error in JavaGroupByBuilder MultiAccumulate

A groupby with multiple accumulators will have accumulators.length + 1
elements in its result tuple: the group key + the result from all the
accumulators. This differs from a plain accumulate, which just has
accumulators.length elements in its result tuple. This commit make all the
groupby tests pass in Java interpreted mode.

* Fix implementation and add tests for MVEL dialect groupby

- Tests are done without duplication by putting the asserter of
  the tested rule into a Consumer<KieSession>, and compiling two
  KieBases that are tested with assertAll:

  . One without "dialect \"mvel\";" added to the beginning (the Java
    dialect asserter)

  . One with "dialect \"mvel\";" added to the beginning (the MVEL
    dialect asserter)

- To make tests compatiable with MVEL, Generics were removed from
  the rules and throw statements are done inside a global consumer.

- The Java groupby and the MVEL groupby key function tuples have different
  requirements:

  . The Java groupby key function tuple must be wrapped inside a tuple
    so indexing will work

  . The MVEL groupby key function tuple must be the original tuple

  to accommodate both requirements, a boolean field is added to
  MVELGroupByAccumulate so the correct tuple is sent to the groupby
  key function.

- groupby's ValueResolver's rule base is not an instance of KnowledgeBaseImpl,
  so use InternalRuleBase (a common supertype between KnowledgeBaseImpl
  and the type of groupby's rule base)  in order to get the main package.

- Mostly copied JavaGroupByBuilder's addGroupByFunctionCompilation to
  MVELGroupByBuilder, making neccessary changes so it makes a
  MVELCompilable instead of a Java Invoker.

- It turns out getUsedDeclarationsIncludingGroupBy was not needed
  and was faulty; used getUsedDeclarations and added the groupby
  declaration to the declaration map instead.

* Re-enable the Java Dialect GroupBy tests

Left it disabled when investigating MVEL and forgot
to re-enable it.